requirements.txt
pandas>=2.2.0
numpy>=1.26.0
pyyaml>=6.0.1
pyarrow>=15.0.0
scikit-learn>=1.4.0
joblib>=1.3.2
jinja2>=3.1.3
matplotlib>=3.8.0
requests>=2.31.0
yfinance>=0.2.40
python-dateutil>=2.9.0
